Extracted get_app_templates
